requirements.txt
flask
supabase
python-dotenv
bcrypt
pandas
gunicorn
openpyxl
